# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KRA, KPI & Task Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/top-performers:
    get:
      operationId: getTopPerformers
      tags: [dashboard]
      summary: Get top performing employees
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top performers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployeePerformance"
...
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /dashboard/employee-summary:
    get:
      operationId: getEmployeeDashboardSummary
      tags: [dashboard]
      summary: Get employee-specific dashboard summary
      parameters:
        - name: employeeId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDashboardSummary"
  /dashboard/pending-approvals:
    get:
      operationId: getPendingApprovals
      tags: [dashboard]
      summary: Get pending KRA closures and task status requests (manager/HOD)
      parameters:
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: approverId
          in: query
          schema:
            type: integer
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Pending approvals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingApprovals"
  /holidays:
    get:
      operationId: listHolidays
      tags: [holidays]
      summary: List all holidays
      responses:
        "200":
          description: List of holidays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Holiday"
    post:
      operationId: createHoliday
      tags: [holidays]
      summary: Create a holiday
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HolidayInput"
      responses:
        "201":
          description: Holiday created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holiday"
  /holidays/{id}:
    delete:
      operationId: deleteHoliday
      tags: [holidays]
      summary: Delete a holiday
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    TaskStatusCount:
      type: object
      required: [status, count]
      properties:
        status: { type: string }
        count:  { type: integer }
    EmployeePerformance:
      type: object
      required: [employeeId, employeeName, departmentName, kpiScore, rating, tasksCompleted]
      properties:
        employeeId:     { type: integer }
        employeeName:   { type: string }
        departmentName: { type: string }
        kpiScore:       { type: number }
        rating:         { type: string }
        tasksCompleted: { type: integer }
    ActivityItem:
      type: object
      required: [id, type, description, actorName, createdAt]
      properties:
        id:          { type: integer }
        type:        { type: string, enum: [task_created, task_completed, task_delayed, kpi_updated, kra_scored, employee_added, kra_submitted, kra_approved, kra_rejected, task_status_requested, task_status_approved, task_status_rejected] }
        description: { type: string }
        actorName:   { type: string }
        entityId:    { type: ["integer", "null"] }
        entityType:  { type: ["string", "null"] }
        createdAt:   { type: string }
...
    EmployeeDashboardSummary:
      type: object
      required: [myTasks, completedTasks, pendingTasks, overdueTasks, myKpiScore, myKpiRating, myKraAvg, taskBreakdown, recentActivity]
      properties:
        myTasks:        { type: integer }
        completedTasks: { type: integer }
        pendingTasks:   { type: integer }
        overdueTasks:   { type: integer }
        myKpiScore:     { type: ["number", "null"] }
        myKpiRating:    { type: ["string", "null"] }
        myKraAvg:       { type: ["number", "null"] }
        taskBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/TaskStatusCount"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
    PendingKraApproval:
      type: object
      required: [id, title, employeeId, employeeName, departmentName, kraStatus, workingHoursElapsed, isOverdue]
      properties:
        id:                  { type: integer }
        title:               { type: string }
        employeeId:          { type: integer }
        employeeName:        { type: string }
        departmentName:      { type: string }
        achievementPct:      { type: ["number", "null"] }
        kraStatus:           { type: string }
        submittedAt:         { type: ["string", "null"] }
        pendingAt:           { type: ["string", "null"] }
        workingHoursElapsed: { type: number }
        isOverdue:           { type: boolean }
    PendingTaskApproval:
      type: object
      required: [id, title, assignedToId, assignedToName, departmentName, status, requestedStatus, workingHoursElapsed, isOverdue]
      properties:
        id:                  { type: integer }
        title:               { type: string }
        assignedToId:        { type: integer }
        assignedToName:      { type: string }
        departmentName:      { type: string }
        status:              { type: string }
        requestedStatus:     { type: string }
        progressPct:         { type: integer }
        pendingAt:           { type: ["string", "null"] }
        workingHoursElapsed: { type: number }
        isOverdue:           { type: boolean }
    HrPendingKraApproval:
      type: object
      required: [id, title, departmentName, createdAt, workingHoursElapsed, isOverdue]
      properties:
        id:                  { type: integer }
        title:               { type: string }
        description:         { type: ["string", "null"] }
        weightage:           { type: number }
        departmentName:      { type: string }
        reviewPeriod:        { type: string }
        createdAt:           { type: string }
        pendingAt:           { type: ["string", "null"] }
        workingHoursElapsed: { type: number }
        isOverdue:           { type: boolean }
    PendingApprovals:
      type: object
      required: [kras, tasks, krasPendingHrApproval]
      properties:
        kras:
          type: array
          items:
            $ref: "#/components/schemas/PendingKraApproval"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/PendingTaskApproval"
        krasPendingHrApproval:
          type: array
          items:
            $ref: "#/components/schemas/HrPendingKraApproval"
    Holiday:
      type: object
      required: [id, date, name, createdAt]
      properties:
        id:        { type: integer }
        date:      { type: string }
        name:      { type: string }
        createdAt: { type: string }
    HolidayInput:
      type: object
      required: [date, name]
      properties:
        date: { type: string }
        name: { type: string }
...