# lib/api-zod/src/generated/types/holidayInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KRA, KPI & Task Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface HolidayInput {
  date: string;
  name: string;
}