# lib/api-zod/src/generated/types/hrPendingKraApproval.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KRA, KPI & Task Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface HrPendingKraApproval {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  weightage?: number;
  departmentName: string;
  reviewPeriod?: string;
  createdAt: string;
  /** @nullable */
  pendingAt?: string | null;
  workingHoursElapsed: number;
  isOverdue: boolean;
}