# lib/api-zod/src/generated/types/pendingKraApproval.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KRA, KPI & Task Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface PendingKraApproval {
  id: number;
  title: string;
  employeeId: number;
  employeeName: string;
  departmentName: string;
  /** @nullable */
  achievementPct?: number | null;
  kraStatus: string;
  /** @nullable */
  submittedAt?: string | null;
  /** @nullable */
  pendingAt?: string | null;
  workingHoursElapsed: number;
  isOverdue: boolean;
}