# lib/api-zod/src/generated/types/pendingTaskApproval.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KRA, KPI & Task Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface PendingTaskApproval {
  id: number;
  title: string;
  assignedToId: number;
  assignedToName: string;
  departmentName: string;
  status: string;
  requestedStatus: string;
  progressPct?: number;
  /** @nullable */
  pendingAt?: string | null;
  workingHoursElapsed: number;
  isOverdue: boolean;
}